Restructure folders for server
