+++ src/server/api/root.ts
@@ -1,0 +1,16 @@
+import { router } from "./trpc";
+import { authRouter } from "./routers/auth";
+import { exampleRouter } from "./routers/example";
+
+/**
+ * This is the primary router for your server.
+ *
+ * All routers added in /api/routers should be manually added here
+ */
+export const appRouter = router({
+  example: exampleRouter,
+  auth: authRouter,
+});
+
+// export type definition of API
+export type AppRouter = typeof appRouter;
