+++ src/server/api/trpc/context.ts
@@ -2,8 +2,8 @@ import { type inferAsyncReturnType } from "@trpc/server";
 import { type CreateNextContextOptions } from "@trpc/server/adapters/next";
 import { type Session } from "next-auth";
 
-import { getServerAuthSession } from "../common/get-server-auth-session";
-import { prisma } from "../db/client";
+import { getServerAuthSession } from "../../common/get-server-auth-session";
+import { prisma } from "../../db/client";
 
 type CreateContextOptions = {
   session: Session | null;
